refactor: extract LLM JSON helpers; fix superpowers path; drop generic
- Move LlmJsonParseError and extractJsonObject out of judge-diff.ts into
  a dedicated packages/corespec/src/llm/json-response.ts module (alongside
  provider.ts), with its own unit tests. Adds parseLlmJson() as the happy-
  path wrapper so judge-diff no longer owns generic JSON-parsing logic.
- Fix superpowers spec directory: obra/superpowers stores plans under
  docs/superpowers/plans/ (not .claude/skills/, which is present for any
  Claude Code plugin). Detection now keys off docs/superpowers/.
- Drop the 'generic' framework entirely — we can't point at a canonical
  spec directory for it, so including it was guessing.
- Add project URLs (OpenSpec, Kiro, Superpowers) as comments next to the
  framework→directory map in format.ts and comment.sh.
- Update pr-comment spec.md to reflect the verified mapping.
