+++ packages/corespec/src/index.ts
@@ -10,8 +10,9 @@ export type {
 
 export { checkFramework } from './tools/check-framework.js';
 export { judgeFramework } from './tools/judge-framework.js';
-export { judgeDiff, LlmJsonParseError } from './tools/judge-diff.js';
+export { judgeDiff } from './tools/judge-diff.js';
 export { loadSpecs } from './tools/load-specs.js';
+export { extractJsonObject, parseLlmJson, LlmJsonParseError } from './llm/json-response.js';
 
 export { parseModelId, callLLM } from './llm/provider.js';
