+++ packages/corespec/src/tools/judge-diff.ts
@@ -1,71 +1,7 @@
 import type { FrameworkJudgment, DiffJudgment, ModelConfig, SpecDocument } from '../types.js';
 import { callLLM } from '../llm/provider.js';
+import { parseLlmJson } from '../llm/json-response.js';
 import picomatch from 'picomatch';
-
-/**
- * Thrown when the LLM returns a response that is not parseable JSON.
- * Carries the raw response and model name so callers can surface actionable
- * error messages instead of bare JSON.parse output.
- */
-export class LlmJsonParseError extends Error {
-  public readonly rawResponse: string;
-  public readonly model: string;
-  public readonly parseError: string;
-
-  constructor(parseError: string, rawResponse: string, model: string) {
-    const preview = rawResponse.trim().slice(0, 200).replace(/\s+/g, ' ');
-    const suffix = rawResponse.length > 200 ? '…' : '';
-    super(
-      `LLM returned a non-JSON response (model: ${model}). ` +
-        `JSON parse error: ${parseError}. ` +
-        `Response preview: "${preview}${suffix}". ` +
-        `The model likely ignored the JSON-only instruction. Try a more capable model ` +
-        `(e.g. claude-sonnet-4) or reduce the prompt size by tightening the ignore list in .specguard.yml.`,
-    );
-    this.name = 'LlmJsonParseError';
-    this.rawResponse = rawResponse;
-    this.model = model;
-    this.parseError = parseError;
-  }
-}
-
-/**
- * Extracts the first balanced JSON object from an LLM response, tolerating
- * leading/trailing prose, markdown code fences, or commentary.
- */
-export function extractJsonObject(raw: string): string {
-  const stripped = raw.replace(/```(?:json)?\s*/g, '').replace(/\s*```/g, '');
-  const start = stripped.indexOf('{');
-  if (start === -1) return stripped.trim();
-
-  let depth = 0;
-  let inString = false;
-  let escape = false;
-  for (let i = start; i < stripped.length; i++) {
-    const ch = stripped[i];
-    if (inString) {
-      if (escape) {
-        escape = false;
-      } else if (ch === '\\') {
-        escape = true;
-      } else if (ch === '"') {
-        inString = false;
-      }
-      continue;
-    }
-    if (ch === '"') {
-      inString = true;
-    } else if (ch === '{') {
-      depth++;
-    } else if (ch === '}') {
-      depth--;
-      if (depth === 0) {
-        return stripped.slice(start, i + 1);
-      }
-    }
-  }
-  return stripped.slice(start).trim();
-}
 
 /** Split a unified diff into per-file hunks and drop files matching ignore patterns. */
 export function filterDiff(diff: string, ignore: string[]): string {
@@ -166,15 +102,10 @@ export async function judgeDiff(
 
   const prompt = buildPrompt(framework, filteredDiff, threshold, specs);
   const raw = await callLLM(config, prompt);
-  const response = extractJsonObject(raw);
-
-  let parsed: { files?: Array<{ file: string; score?: number; reason?: string; gap?: string }>; threshold?: number };
-  try {
-    parsed = JSON.parse(response);
-  } catch (err) {
-    const parseMessage = err instanceof Error ? err.message : String(err);
-    throw new LlmJsonParseError(parseMessage, raw, config.model);
-  }
+  const parsed = parseLlmJson<{
+    files?: Array<{ file: string; score?: number; reason?: string; gap?: string }>;
+    threshold?: number;
+  }>(raw, config.model);
 
   const files = (parsed.files ?? []).map(f => ({
     file: f.file,
